Corrected CTDI calibration positioning

+++ src/dxmclib/src/source.cpp
@@ -31,7 +31,7 @@ constexpr double DEG_TO_RAD = PI / 180.0;
 constexpr double RAD_TO_DEG = 1.0 / DEG_TO_RAD;
 constexpr double KEV_TO_MJ = 1.6021773e-13;
 
-constexpr std::uint64_t CTDI_MIN_HISTORIES = 500E6;
+constexpr std::uint64_t CTDI_MIN_HISTORIES = static_cast<std::uint64_t>(500E6);
 
 template<typename T>
 T vectorLenght(const std::array<T, 3>& v)
